feat: create bookkeeper proxy model
Bookkeeper model move methods and attributes to new proxy model
