+++ src/bookkeeper_checklist_project/bookkeeper/models/__init__.py
@@ -1,1 +1,2 @@
 from .bookkeeper import Bookkeeper, get_all_bookkeepers_as_choices
+from .bookkeeper_proxy import BookkeeperProxy
